requirements.txt
discord
pynacl
google-cloud-texttospeech
gtts
python-dotenv
yt_dlp
ffmpeg
pydub
demucs
ipython
so-vits-svc-fork
dacite
openai
PyYAML